chore: remove cargo2nix
This commit changes from cargo2nix to just use `buildRustPackage`, which
does not generate itermediate crate derivations for the build. This
means that we lose dependency caching (for CI as well), but on the other
hand it becomes much more cacheable for people that actually use Nix to
add the dependencies.