Initial pass at watch filtering
This allows you to type and hit enter while running check --watch,
which will update the -m setting, basically.

This can be super useful for filtering down to just one test to run and
iterate on, before pressing enter again to clear it out and go back to
running everything again.