Switch to a png favicon in generated docs
The aiken stdlib favicon has been broken for me for ages, which causes it to disappear into my tab list, and it's been driving me nuts; this is because the SVG used has a large width and height, and so the browser is rendering a very tiny transparent rectangle from the top left corner. You can confirm this by replacing the SVG with a full black rectangle. Since the favicon on the main aiken site works just fine, this just uses that.