Make UPLC Term generic over context
Adds a generic "Context" parameter that can be populated with data; any
methods which work with terms will preserve the context, meaning it
survives execution, optimization, etc.

For now, we default it to (), but this is foundational work for source
map support, because we can assign sourcemap locations as the context
during codegen.