Make optimization pass generic over context
It doesn't do us much good to carry the source spans through the whole
compilation process, if optimizations just screw with the UPLC tree at
the end.

This ensures that the interning, shrinking, and other optimizations
carry through the relevant context.

This also fixes the order of operations to apply used functions before
optimization.